Finished the Rust DFA implementation. Started NFA.
The rust implementation is done, and you can read the README file
to get a sense of what's different between it and Haskell.  The
Haskell NFA has been started, but I'm still working on learning the
ins and outs of list zipping and flattening, which are crucial for
this particular project.